haskellPackages.spago: rename to `spago-legacy`
The Spago package in NixPkgs has moved from
https://github.com/purescript/spago to https://github.com/purescript/spago-legacy.

I believe the new `spago` https://github.com/purescript/spago is a rewrite from Haskell to PureScript. To support this
new Spago in NixPkgs will probably require rewriting all of the Nix configuration. I.m.o. the easiest way to
clear the path for this is to rename the old Nix configuration from `spago` to `spago-legacy`.

Because part of this is generated by `cabal2nix`, changes to the `spago-legacy` Cabal file are needed: https://github.com/purescript/spago-legacy/pull/3